create Note component to hold all note content and children
