+++ components/gameEntry/GameEntry.module.scss
@@ -39,35 +39,4 @@
       min-width: fit-content;
     }
   }
-  
-  // third child
-  & > div:nth-child(3) {
-    padding-inline: 1rem;
-
-    hr {
-      margin-block: 0.25rem;
-      border-color: hsl(275, 100%, 66%);
-    }
-
-    & > p {
-      display: flex;
-      justify-content: space-between;
-
-      h4 {
-        text-decoration: underline;
-        font-size: 1.05rem;
-        font-weight: 500;
-      }
-    }
-  }
-
-  button {
-    // background-color: var(--bkgd-btn-edit);
-    // color: var(--text-btn-edit);
-    // border-radius: var(--bdr-rad-btn);
-    // padding: var(--pad-btn);
-    // min-width: fit-content;
-  }
-  
-
 }
